Apply suggestions from code review

+++ tests/test_id3v2.cpp
@@ -1746,13 +1746,13 @@ public:
                                                   "\xFF\x00"
                                                   "\x05"
                                                   "\x14\x4F\x00\x32", 10));
-    CPPUNIT_ASSERT_EQUAL(invalidVersionHeader.tagSize(), static_cast<unsigned int>(0));
+    CPPUNIT_ASSERT_EQUAL(invalidVersionHeader.tagSize(), 0U);
 
     ID3v2::Header invalidRevisionHeader(ByteVector("ID3"
                                                    "\x04\xFF"
                                                    "\x05"
                                                    "\x14\x4F\x00\x32", 10));
-    CPPUNIT_ASSERT_EQUAL(invalidRevisionHeader.tagSize(), static_cast<unsigned int>(0));
+    CPPUNIT_ASSERT_EQUAL(invalidRevisionHeader.tagSize(), 0U);
   }
 
 };
